+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add a command for copying links to clipboard.
 - Add chafa as fallback for image rendering.
 
+### Changed
+- Use Inline Images Protocol for mlterm.
+
 ### Fixed
 - Handle cell sizes with floating point numbers.
 - Don't divide by zero for fallback cell size detection.